MAINT: Move find_prs.py into find_pr_api
Moving the find_prs.py module into the find_pr_api directory. Renaming the module to github.py to show that this module finds github prs. This directory will contain other modules to find requests from services such as GitLab and Phabricator
