+++ cloud-chatops/src/events.py
@@ -10,7 +10,7 @@ import schedule
 
 from data import User
 from slack_reminder_api.pr_reminder import PRReminder
-from find_prs import FindPRs
+from find_pr_api.github import FindPRs
 from read_data import get_config, get_token
